backend/requirements.txt
fastapi
uvicorn
requests
pydantic
duckduckgo-search
sentence-transformers
faiss-cpu